REF: Rename `openfast` package to `openfast_service`
